refactor: use only one screen reader
