+++ apis/nucleus/src/components/listbox/ListBox.jsx
@@ -14,7 +14,7 @@ import ListBoxFooter from './components/ListBoxFooter';
 import getScrollIndex from './interactions/listbox-get-scroll-index';
 import getFrequencyAllowed from './components/grid-list-components/frequency-allowed';
 import useFrequencyMax from './hooks/useFrequencyMax';
-import { ScreenReaderForSelections, getScreenReaderSearchText } from './components/ScreenReaders';
+import getScreenReaderAssertiveText from './components/screen-reader/assertive-screen-reader';
 import InstanceContext from '../../contexts/InstanceContext';
 import deduceFrequencyMode from './utils/deduce-frequency-mode';
 
@@ -60,7 +60,6 @@ export default function ListBox({
   const [initScrollPosIsSet, setInitScrollPosIsSet] = useState(false);
   const isSingleSelect = !!(layout && layout.qListObject.qDimensionInfo.qIsOneAndOnlyOne);
   const { checkboxes = checkboxOption, histogram } = layout ?? {};
-  const [screenReaderText, setScreenReaderText] = useState('');
 
   const loaderRef = useRef(null);
   const local = useRef({
@@ -218,15 +217,8 @@ export default function ListBox({
   const { listCount } = sizes;
   setStoreValue('listCount', listCount);
 
-  const inputText = getStoreValue('inputText');
-
-  useEffect(() => {
-    if (inputText) {
-      const srText = getScreenReaderSearchText(listCount);
-      const srFinalText = translatorDynamic.get(srText, [listCount]);
-      setScreenReaderText(srFinalText);
-    }
-  }, [inputText, listCount]);
+  const searchInputText = getStoreValue('inputText');
+  const screenReaderText = getScreenReaderAssertiveText({ layout, searchInputText, listCount });
 
   const setScrollPosition = (position) => {
     const { scrollIndex, offset, triggerRerender } = getScrollIndex({
@@ -317,7 +309,6 @@ export default function ListBox({
 
   return (
     <StyledWrapper>
-      <ScreenReaderForSelections className="screenReaderOnly" layout={layout} />
       <div className="screenReaderOnly" aria-live="assertive">
         {screenReaderText}
       </div>
